requirements.txt
pandas~=1.3.5
numpy~=1.21.6
pyomo~=6.6.2
matplotlib~=3.5.3
scipy~=1.7.3
pvlib~=0.10.2
platypus-optwindpowerlib~=1.1.0